requirements-recolorization.txt
pillow
argparse
numpy
scikit-image
scikit-learn
sklearn
typing-extensions
opencv-python
torch
torchvision
dominate
visdom
